requirements.txt
django-celery-beat
celery[redis]
djangorestframework
django
celery
pytest-django
python-dotenv
